setup/requirements.txt
kafka-python==2.0.2
gqlalchemy==1.0.5
